🎉 팔로우 토글 버튼, 팔로워/팔로잉 목록 생성

+++ src/hooks/useFollow.ts
@@ -5,7 +5,6 @@ import { deleteFollow, postFollow } from '@/services/follow'
 import User from '@/types/user'
 import { useCurrentUser } from './useCurrentUser'
 
-//FIXME: 정리되지 않은 코드입니다. 정리 후 사용해주세요.
 /**
  *
  * @returns param isFollowing: 로그인한 유저가 현재 페이지의 유저를 팔로우하고 있는지 여부
@@ -15,7 +14,7 @@ import { useCurrentUser } from './useCurrentUser'
  * @returns param followerCount: 현재 페이지의 유저의 팔로워 수
  * @returns param followingCount: 현재 페이지의 유저의 팔로잉 수
  */
-const useFollow = (userData: User) => {
+const useFollow = (userData: User | undefined) => {
   const { currentUser } = useCurrentUser()
 
   const [isFollowing, setIsFollowing] = useState(false)
@@ -25,7 +24,7 @@ const useFollow = (userData: User) => {
   const [followingCount, setFollowingCount] = useState(0)
 
   useEffect(() => {
-    if (!currentUser) return
+    if (!userData || !currentUser) return
     setFollowerCount(() => userData.followers?.length ?? 0)
     setFollowingCount(() => userData.following?.length ?? 0)
     setIsFollowing(
@@ -40,24 +39,19 @@ const useFollow = (userData: User) => {
           (follower) => userData._id === follower._id,
         ) ?? false,
     )
-  }, [
-    currentUser,
-    userData._id,
-    userData.followers,
-    userData.following?.length,
-  ])
+  }, [currentUser, userData])
 
   const followToggle = async () => {
-    if (!currentUser) return
+    if (!userData || !currentUser) return
     if (isFollowing) {
+      setFollowerCount((prev) => prev - 1)
+      setIsFollowing(false)
       const followData = currentUser.following?.find(
         ({ user }) => user === userData._id,
       )
-      console.log(userData, currentUser, followData)
 
       setFollowId(() => followData?._id ?? '')
       await deleteFollow(followData?._id ?? followId)
-      setIsFollowing(false)
 
       currentUser.following?.splice(
         currentUser.following.findIndex(
@@ -71,10 +65,10 @@ const useFollow = (userData: User) => {
         ),
         1,
       )
-      setFollowerCount((prev) => prev - 1)
     } else {
-      const followData = await postFollow(userData._id)
       setIsFollowing(true)
+      setFollowerCount((prev) => prev + 1)
+      const followData = await postFollow(userData._id)
       setFollowId(() => followData._id ?? '')
       userData.followers?.push({
         _id: currentUser._id,
@@ -84,7 +78,6 @@ const useFollow = (userData: User) => {
         updatedAt: new Date().toISOString(),
       })
       currentUser.following?.push(followData)
-      setFollowerCount((prev) => prev + 1)
     }
   }
 
@@ -93,7 +86,7 @@ const useFollow = (userData: User) => {
     isFollowed,
     followerCount,
     followingCount,
-    unavailable: !currentUser || currentUser._id === userData._id,
+    unavailable: !currentUser || currentUser._id === userData?._id,
     followToggle,
   }
 }
